Added light pole rotation
Added light pole rotation with little or no change to any of light
making code. This gives me the ability to also change sizes as well with
no ability of disrupting current code.

next change will be to the roads then actually starting work on the
timer and math for checking what lights are on

after that is adding a visual of cars going and stopping at the lights.